Refactor the authenticating state
