+++ packages/react/src/components/form/authenticating/authenticating.components.tsx
@@ -1,0 +1,40 @@
+import { LinkButton } from "../../../../../react-primitives/src/components/button/link-button";
+import { sprinkles } from "../../../../../react-primitives/src/theme/sprinkles.css";
+import { useConfiguration } from "../../../hooks/use-configuration";
+import { Text } from "../../text";
+
+import * as styles from "./authenticating.css";
+
+export const BackButton = ({ onCancel }: { onCancel: () => void }) => {
+  const { text } = useConfiguration();
+  return (
+    <LinkButton
+      className={sprinkles({ marginBottom: "4" })}
+      testId="sid-form-authenticating-cancel-button"
+      variant="back"
+      onClick={onCancel}
+    >
+      {text["authenticating.back"]}
+    </LinkButton>
+  );
+};
+
+export const RetryPrompt = ({ onRetry }: { onRetry: () => void }) => {
+  const { text } = useConfiguration();
+  return (
+    <div className={styles.retryPrompt}>
+      <Text
+        variant={{ size: "sm", color: "tertiary", weight: "semibold" }}
+        t="authenticating.retryPrompt"
+      />
+      <LinkButton
+        className={sprinkles({ marginLeft: "1" })}
+        type="button"
+        testId="sid-form-authenticating-retry-button"
+        onClick={onRetry}
+      >
+        {text["authenticating.retry"]}
+      </LinkButton>
+    </div>
+  );
+};
