+++ packages/react/src/components/form/authenticating/authenticating.types.ts
@@ -1,0 +1,5 @@
+import { AuthenticatingState } from "../flow";
+
+export type Props = {
+  flowState: AuthenticatingState;
+};
