+++ packages/react/src/components/form/authenticating/password.tsx
@@ -1,0 +1,7 @@
+import { Props } from "./authenticating.types";
+
+// special case handling for password fields
+export const PasswordState = (props: Props) => {
+  console.log({ props });
+  return <div>password</div>;
+};
