+++ packages/react/src/dev.tsx
@@ -292,8 +292,9 @@ root.render(
     <SlashIDProvider
       oid={import.meta.env.VITE_ORG_ID}
       themeProps={{ theme: "dark" }}
-      baseApiUrl="https://api.sandbox.slashid.com"
-      tokenStorage="localStorage"
+      baseApiUrl="https://slashid.local"
+      sdkUrl="https://jump.slashid.local/sdk.html"
+      tokenStorage="memory"
       analyticsEnabled
     >
       <LogOut />
