+++ packages/react/src/domain/types.ts
@@ -1,11 +1,8 @@
 import { Factor, SlashID, User } from "@slashid/slashid";
 import { ReactNode } from "react";
 
-const handleTypes = [
-  "email_address",
-  "phone_number"
-] as const
-export type HandleType = typeof handleTypes[number]
+const handleTypes = ["email_address", "phone_number"] as const;
+export type HandleType = (typeof handleTypes)[number];
 
 export interface Handle {
   type: HandleType;
@@ -13,15 +10,15 @@ export interface Handle {
 }
 
 export const isHandle = (obj: unknown): obj is Handle => {
-  if (!obj) return false
-  if (Array.isArray(obj)) return false
-  if (typeof obj !== "object") return false
+  if (!obj) return false;
+  if (Array.isArray(obj)) return false;
+  if (typeof obj !== "object") return false;
 
-  const { type, value } = obj as Handle
-  if (!handleTypes.includes(type)) return false
-  
-  return typeof value === "string"
-}
+  const { type, value } = obj as Handle;
+  if (!handleTypes.includes(type)) return false;
+
+  return typeof value === "string";
+};
 
 export interface LoginConfiguration {
   handle?: Handle;
@@ -50,6 +47,8 @@ export type FactorNonOIDC = Exclude<
   FactorOIDC | FactorLabeledOIDC
 >;
 export type FactorSSO = FactorLabeledOIDC | FactorCustomizableSAML;
+
+export type FactorPassword = Extract<Factor, { method: "password" }>;
 
 export type FactorNonSSO = Exclude<FactorConfiguration, FactorSSO>;
